Keep detectors.py's shape and put the new capability in new files
The reviewer kept describing detectors.py as it exists on main, and the count
of findings moved every time code moved inside that file, which is the tell:
its diff was too large to read. Gutting the module for the earlier split made
that worse, not better, because 1,527 deleted lines are what the reviewer had
to wade through to reach 100 added ones.

So the shape is inverted. detectors.py now starts from what main has and
changes surgically: the action surface on each normalized step, thresholds
threaded through the four trajectory detectors, the registry, run_all,
session_profile. Everything genuinely new is in a module of its own, which is
also where a reader would look for it:

  detector_surface.py     what a call touched, and what a finding may repeat
  detector_calibration.py where a threshold comes from
  detector_behaviour.py   the four behavioural detectors
  detector_money.py       what a finding costs

detectors.py's diff goes from 851 lines to 421, and session_profile, which
both remaining findings were about, moves from line 816 of that diff to line
132. detector_core.py is gone; the code that was always in detectors.py is
back in it, which is the half of the earlier split that was wrong.

detector_behaviour late-imports three primitives from the orchestrator that
imports it, the shape this repo already uses for a leaf that needs something
from its parent. All 112 tests pass with no test changed. CLAUDE.md and the
blueprint's module table both follow the code.
